refactor(keystone-operator): move watch mappers to keystone_watches.go
The four Secret/MariaDB/ClusterSecretStore/PushSecret event mappers,
the PushSecret change predicate, and the finalizersEqualAsSet helper
(~250 lines of watch plumbing) lived in keystone_controller.go and
crowded out the reconcile chain. Extract them verbatim into a new
keystone_watches.go in the same package so the controller file stays
focused on Reconcile() and SetupWithManager.

Pure code move: no behaviour change, no exported-symbol change. The
now-unused slices/types/event/predicate/reconcile imports are dropped
from keystone_controller.go and declared in keystone_watches.go.

Part of issue #467.

AI-assisted: Claude Code
On-behalf-of: @SAP christian.berendt@sap.com
